Move keras implementation functions into PruningKerasImplementation
