+++ model_compression_toolkit/core/common/pruning/memory_calculator.py
@@ -367,7 +367,7 @@ class MemoryCalculator:
             int: The total number of parameters in the node after pruning.
         """
         total_params = 0
-        attributes_and_oc_axis = self.fw_impl.get_node_attributes_with_io_axis(node, self.fw_info)
+        attributes_and_oc_axis = self.fw_impl.get_node_attributes_with_oi_axis(node, self.fw_info)
 
         # Iterate over the node's weights and apply pruning based on the masks.
         for w_attr, w in node.weights.items():
