+++ model_compression_toolkit/core/common/pruning/pruning_framework_implementation.py
@@ -13,7 +13,7 @@
 # limitations under the License.
 # ==============================================================================
 
-from typing import List, Tuple
+from typing import List, Tuple, Dict
 
 from abc import abstractmethod
 
@@ -105,20 +105,9 @@ class PruningFrameworkImplementation(FrameworkImplementation):
     @abstractmethod
     def is_node_exit_node(self,
                           node: BaseNode,
-                          dual_entry_node: BaseNode):
-        """
-        Abstract method to determine if a given node is an exit node.
+                          corresponding_entry_node: BaseNode,
+                          fw_info: FrameworkInfo):
 
-        Args:
-            node: The node to be checked.
-            dual_entry_node: Another node to be used in the determination process.
-
-        Returns:
-            bool: True if the node is an exit node, False otherwise.
-
-        Raises:
-            NotImplemented: If the method is not implemented in the subclass.
-        """
         raise NotImplemented(f'{self.__class__.__name__} have to implement the '
                              f'framework\'s is_node_exit_node method.')  # pragma: no cover
 
@@ -140,17 +129,19 @@ class PruningFrameworkImplementation(FrameworkImplementation):
         raise NotImplemented(f'{self.__class__.__name__} have to implement the '
                              f'framework\'s is_node_intermediate_pruning_section method.')  # pragma: no cover
 
-    def get_node_attributes_with_io_axis(self, node: BaseNode, fw_info: FrameworkInfo):
+    def get_node_attributes_with_oi_axis(self, node: BaseNode, fw_info: FrameworkInfo) -> Dict[str, Tuple[int, int]]:
         """
-        Gets the attributes of a node and the axis for each attribute's output channels dimension.
+        Gets the attributes of a node and the axis for each attribute's output/input
+        channels dimension.
 
         Args:
-            node (BaseNode): The node for which attributes and their output channel axis are required.
-            fw_info (FrameworkInfo): Framework-specific information containing details about layers and attributes.
+            node (BaseNode): The node for which attributes and their oi channel axis are required.
+            fw_info (FrameworkInfo): Framework-specific information containing details about
+            layers and attributes.
 
         Returns:
-            List[Tuple[str, int]]: A list of tuples where each tuple contains an attribute name and the axis
-                                   of the output channels for that attribute.
+            Dict[str, Tuple[int, int]]: A dict of the node's attributes their oi axis.
+
         """
         raise NotImplemented(f'{self.__class__.__name__} have to implement the '
                              f'framework\'s get_node_attributes_with_output_axis method.')  # pragma: no cover
