+++ model_compression_toolkit/core/keras/pruning/pruning_keras_implementation.py
@@ -13,20 +13,19 @@
 # limitations under the License.
 # ==============================================================================
 
-from typing import List, Tuple
+from typing import List, Tuple, Dict
 
-from model_compression_toolkit.core.common.framework_info import FrameworkInfo
-from model_compression_toolkit.core.common import BaseNode
 from model_compression_toolkit.core.common.pruning.pruning_framework_implementation import \
     PruningFrameworkImplementation
+from model_compression_toolkit.core.common.pruning.pruning_section import PruningSection
 from model_compression_toolkit.core.keras.keras_implementation import KerasImplementation
-from model_compression_toolkit.core.keras.pruning.attributes_info import get_keras_node_attributes_with_io_axis
-from model_compression_toolkit.core.keras.pruning.check_node_role import is_keras_node_intermediate_pruning_section, \
-    is_keras_entry_node, is_keras_exit_node
-from model_compression_toolkit.core.keras.pruning.prune_keras_node import (prune_keras_exit_node,
-                                                                           prune_keras_entry_node, \
-                                                                           prune_keras_intermediate_node)
+from model_compression_toolkit.core.common.framework_info import FrameworkInfo
+from model_compression_toolkit.core.common import BaseNode
+from model_compression_toolkit.core.keras.constants import BIAS, GROUPS, FILTERS, UNITS, USE_BIAS
+import keras
+
 import numpy as np
+
 
 class PruningKerasImplementation(KerasImplementation, PruningFrameworkImplementation):
     """
@@ -35,7 +34,10 @@ class PruningKerasImplementation(KerasImplementation, PruningFrameworkImplementa
     for the Keras framework.
     """
 
-    def prune_entry_node(self, node: BaseNode, output_mask: np.ndarray, fw_info: FrameworkInfo):
+    def prune_entry_node(self,
+                         node: BaseNode,
+                         output_mask: np.ndarray,
+                         fw_info: FrameworkInfo):
         """
         Prunes the entry node of a model in Keras.
 
@@ -44,12 +46,17 @@ class PruningKerasImplementation(KerasImplementation, PruningFrameworkImplementa
             output_mask: A numpy array representing the mask to be applied to the output channels.
             fw_info: Framework-specific information object.
 
-        Returns:
-            The result from the pruning operation.
         """
-        return prune_keras_entry_node(node, output_mask, fw_info)
+        return _prune_keras_edge_node(node=node,
+                                      mask=output_mask,
+                                      fw_info=fw_info,
+                                      is_exit_node=False)
 
-    def prune_intermediate_node(self, node: BaseNode, input_mask: np.ndarray, output_mask: np.ndarray, fw_info: FrameworkInfo):
+    def prune_intermediate_node(self,
+                                node: BaseNode,
+                                input_mask: np.ndarray,
+                                output_mask: np.ndarray,
+                                fw_info: FrameworkInfo):
         """
         Prunes an intermediate node in a Keras model.
 
@@ -59,12 +66,19 @@ class PruningKerasImplementation(KerasImplementation, PruningFrameworkImplementa
             output_mask: A numpy array representing the mask to be applied to the output channels.
             fw_info: Framework-specific information object.
 
-        Returns:
-            The result from the pruning operation.
         """
-        return prune_keras_intermediate_node(node, input_mask, output_mask, fw_info)
+        _edit_node_input_shape(input_mask, node)
+        pruned_parameters = {}
+        mask_bool = output_mask.astype(bool)
+        for k, v in node.weights.items():
+            # Apply the mask to the weights.
+            pruned_parameters[k] = v.compress(mask_bool, axis=-1)
+        node.weights = pruned_parameters
 
-    def prune_exit_node(self, node: BaseNode, input_mask: np.ndarray, fw_info: FrameworkInfo):
+    def prune_exit_node(self,
+                        node: BaseNode,
+                        input_mask: np.ndarray,
+                        fw_info: FrameworkInfo):
         """
         Prunes the exit node of a model in Keras.
 
@@ -73,12 +87,13 @@ class PruningKerasImplementation(KerasImplementation, PruningFrameworkImplementa
             input_mask: A numpy array representing the mask to be applied to the input channels.
             fw_info: Framework-specific information object.
 
-        Returns:
-            The result from the pruning operation.
         """
-        return prune_keras_exit_node(node, input_mask, fw_info)
+        return _prune_keras_edge_node(node=node,
+                                      mask=input_mask,
+                                      fw_info=fw_info,
+                                      is_exit_node=True)
 
-    def is_node_entry_node(self, node: BaseNode):
+    def is_node_entry_node(self, node: BaseNode) -> bool:
         """
         Determines whether a node is an entry node in a Keras model.
 
@@ -88,22 +103,28 @@ class PruningKerasImplementation(KerasImplementation, PruningFrameworkImplementa
         Returns:
             Boolean indicating if the node is an entry node.
         """
-        return is_keras_entry_node(node)
+        return _is_keras_node_pruning_section_edge(node)
 
-    def is_node_exit_node(self, node: BaseNode, corresponding_entry_node: BaseNode, fw_info: FrameworkInfo):
+    def is_node_exit_node(self,
+                          node: BaseNode,
+                          corresponding_entry_node: BaseNode,
+                          fw_info: FrameworkInfo) -> bool:
         """
         Determines whether a node is an exit node in a Keras model.
 
         Args:
             node: The node to be checked.
-            corresponding_entry_node: A related entry node to assist in the determination.
+            corresponding_entry_node: The entry node of the pruning section that is checked.
+            fw_info: Framework-specific information object.
 
         Returns:
             Boolean indicating if the node is an exit node.
         """
-        return is_keras_exit_node(node, corresponding_entry_node, fw_info)
+        return _is_keras_node_pruning_section_edge(node) and PruningSection.has_matching_channel_count(node,
+                                                                                                       corresponding_entry_node,
+                                                                                                       fw_info)
 
-    def is_node_intermediate_pruning_section(self, node):
+    def is_node_intermediate_pruning_section(self, node) -> bool:
         """
         Determines whether a node is part of the intermediate section in the pruning process of a Keras model.
 
@@ -113,20 +134,138 @@ class PruningKerasImplementation(KerasImplementation, PruningFrameworkImplementa
         Returns:
             Boolean indicating if the node is part of the intermediate pruning section.
         """
-        return is_keras_node_intermediate_pruning_section(node)
+        # Nodes that are not Conv2D, Conv2DTranspose, DepthwiseConv2D, or Dense are considered intermediate.
+        return node.type not in [keras.layers.DepthwiseConv2D,
+                                 keras.layers.Conv2D,
+                                 keras.layers.Conv2DTranspose,
+                                 keras.layers.Dense]
 
-    def get_node_attributes_with_io_axis(self,
-                                             node: BaseNode,
-                                             fw_info: FrameworkInfo) -> List[Tuple[str, int]]:
+    def get_node_attributes_with_oi_axis(self,
+                                         node: BaseNode,
+                                         fw_info: FrameworkInfo) -> Dict[str, Tuple[int, int]]:
         """
-        Gets the attributes of a node and the axis for each attribute's output channels dimension.
+        Retrieves the attributes of a given node along with the output/input (OI) channel axis
+        for each attribute.
+
+        For kernel operations (like convolutions), the function identifies the output and input
+        channel axis based on framework-specific information.
+        For non-kernel operations, it defaults to setting the last axis as the output
+        channel axis, assuming no specific input channel axis.
 
         Args:
-            node (BaseNode): The node for which attributes and their output channel axis are required.
-            fw_info (FrameworkInfo): Framework-specific information containing details about layers and attributes.
+            node (BaseNode): The node from the computational graph.
+            fw_info (FrameworkInfo): Contains framework-specific information and utilities.
 
         Returns:
-            List[Tuple[str, int]]: A list of tuples where each tuple contains an attribute name and the axis
-                                   of the output channels for that attribute.
+            Dict[str, Tuple[int, int]]: A dictionary where each key is an attribute name (like 'kernel' or 'bias')
+            and each value is a tuple representing the output and input channel axis indices respectively.
         """
-        return get_keras_node_attributes_with_io_axis(node, fw_info)
+
+        attributes_with_axis = {}
+        if fw_info.is_kernel_op(node.type):
+            kernel_attributes = fw_info.get_kernel_op_attributes(node.type)
+            for attr in kernel_attributes:
+                attributes_with_axis[attr] = fw_info.kernel_channels_mapping.get(node.type)
+
+            # Bias is usually associated with output channels, so the input axis is None.
+            attributes_with_axis['bias'] = (0, None)
+        else:
+            # For non-kernel operations, iterate over the node's weights.
+            for attr in list(node.weights.keys()):
+                # The output channel axis is assumed to be the last axis, with no specific input axis.
+                attributes_with_axis[attr] = (-1, None)
+
+        return attributes_with_axis
+
+
+def _is_keras_node_pruning_section_edge(node: BaseNode) -> bool:
+    """
+    Determines if a Keras node is an edge of a pruning section.
+
+    In the context of pruning, an 'edge' node is a layer that can potentially be pruned.
+    This function identifies such nodes based on their type and attributes. Specifically,
+    Conv2D and Conv2DTranspose layers with 'groups' attribute set to 1, and Dense layers
+    are considered as edges for pruning sections.
+
+    Args:
+        node (BaseNode): The node to be evaluated.
+
+    Returns:
+        bool: True if the node is an edge of a pruning section, False otherwise.
+    """
+
+    # Check if the node is a Conv2D or Conv2DTranspose layer with groups set to 1.
+    if node.type in [keras.layers.Conv2D, keras.layers.Conv2DTranspose]:
+        return node.framework_attr[GROUPS] == 1
+    return node.type == keras.layers.Dense
+
+
+
+def _prune_keras_edge_node(node: BaseNode,
+                           mask: np.ndarray,
+                           fw_info: FrameworkInfo,
+                           is_exit_node: bool):
+    """
+    Prunes the given Keras node by applying the mask to the node's weights (kernels and biases).
+    This function can handle both entry and exit nodes by specifying the is_exit_node parameter.
+
+    Args:
+        node: The node to be pruned.
+        mask: The pruning mask to be applied.
+        fw_info: Framework-specific information object.
+        is_exit_node: A boolean indicating whether the node is an exit node.
+
+    """
+
+    # Retrieve the kernel attribute and the axes to prune.
+    kernel_attr = fw_info.get_kernel_op_attributes(node.type)[0]
+    io_axis = fw_info.kernel_channels_mapping.get(node.type)
+    axis_to_prune = io_axis[int(is_exit_node)]
+    kernel = node.get_weights_by_keys(kernel_attr)
+    # Convert mask to boolean.
+    mask_bool = mask.astype(bool)
+
+    pruned_kernel = kernel.compress(mask_bool, axis=axis_to_prune)
+    node.set_weights_by_keys(name=kernel_attr, tensor=pruned_kernel)
+
+    if not is_exit_node and node.framework_attr[USE_BIAS]:
+        # Prune the bias if applicable and it's an entry node.
+        bias = node.get_weights_by_keys(BIAS)
+        pruned_bias = bias.compress(mask_bool)
+        node.set_weights_by_keys(name=BIAS, tensor=pruned_bias)
+
+    if not is_exit_node:
+        # Update 'filters' or 'units' attributes for entry node Conv2D/Conv2DTranspose layers.
+        if node.type in [keras.layers.Conv2D, keras.layers.Conv2DTranspose]:
+            node.framework_attr[FILTERS] = int(np.sum(mask))
+        elif node.type == keras.layers.Dense:
+            node.framework_attr[UNITS] = int(np.sum(mask))
+
+    if is_exit_node:
+        # Adjust the input shape for the last node in the section.
+        _edit_node_input_shape(mask_bool, node)
+
+
+def _edit_node_input_shape(input_mask: np.ndarray,
+                           node: BaseNode):
+    """
+    Adjusts the input shape of a node based on the given input mask.
+
+    This function modifies the input shape of the given node to reflect the pruning
+    that has taken place. It updates the last dimension of the node's input shape
+    to match the number of channels that remain after pruning.
+
+    Args:
+        input_mask (np.ndarray): A binary array where 1 indicates the channel is kept and 0 means pruned.
+        node (BaseNode): The node whose input shape needs to be adjusted.
+    """
+    # Start with the current input shape of the node.
+    new_input_shape = list(node.input_shape)
+
+    # Adjust the last dimension of the shape to match the number of unpruned (retained) channels.
+    # This is done by summing the mask, as each '1' in the mask represents a retained channel.
+    new_input_shape[-1] = int(np.sum(input_mask))
+
+    # Update the node's input shape with the new dimensions.
+    node.input_shape = tuple(new_input_shape)
+
